+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+- add DuckDB::Result#_to_decimal_internal
+- add DuckDB::Result#_to_hugeint_internal
+
+## Breaking Change
+- DuckDB::Result returns BigDecimal object instead of String object if the column type is DECIMAL.
+
 # 0.7.1
 - bump duckdb to 0.7.1
 - fix docker build error on M1 Mac